+++ app/Tests/CourseControllerTests.hs
@@ -1,43 +1,22 @@
-module RequirementTests.CourseControllerTests
+{-|
+Description: Course Controller Module Tests.
+
+Module that contains the tests for the functions in the Course Controller Module.
+
+-}
+
+module Tests.CourseControllerTests
 ( courseContTestSuite ) where
 
 import Config (runDb)
-import Data.Text as T
-import qualified Data.ByteString.Lazy.Char8 as BL
-import qualified Data.Map as Map
-import Test.HUnit
-import Happstack.Server (HttpVersion(..), Request(..), simpleHTTP'', rsBody, Response, ServerPart, Method(GET))
-import Control.Concurrent.MVar (newMVar, newEmptyMVar)
-import Database.Database (clearDatabase)
-import Database.Persist.Sqlite (SqlPersistM, insert_)
-import Database.Tables
 import Controllers.Course (index)
-
--- | A minimal mock request for running a ServerPart
-mockRequest :: IO Request
-mockRequest = do
-    inputsBody <- newMVar []
-    requestBody <- newEmptyMVar
-    return Request
-        { rqSecure          = False             
-        , rqMethod          = GET               
-        , rqPaths           = []                
-        , rqUri             = "/"               
-        , rqQuery           = ""                
-        , rqInputsQuery     = []
-        , rqInputsBody      = inputsBody
-        , rqCookies         = []                
-        , rqVersion         = HttpVersion 1 1
-        , rqHeaders         = Map.empty
-        , rqBody            = requestBody
-        , rqPeer            = ("127.0.0.1", 0)  
-        }
-
--- | Helper function to run ServerPart Response
-runServerPart :: ServerPart Response -> IO Response
-runServerPart sp = do
-    request <- mockRequest
-    simpleHTTP'' sp request
+import qualified Data.ByteString.Lazy.Char8 as BL
+import Data.Text as T
+import Database.Persist.Sqlite (SqlPersistM, insert_)
+import Database.Tables (Courses(..))
+import Happstack.Server (rsBody)
+import Test.HUnit (Test(..), assertEqual)
+import Tests.TestHelpers (clearDatabase, runServerPart)
 
 -- | Helper function to insert courses into the database
 insertCourses :: [T.Text] -> SqlPersistM ()
@@ -58,8 +37,9 @@ indexTestCases =
 runIndexTest :: String -> [T.Text] -> String -> Test
 runIndexTest label courses expected =
     TestLabel label $ TestCase $ do
-        runDb clearDatabase
-        runDb $ insertCourses courses
+        runDb $ do
+            clearDatabase
+            insertCourses courses
         response <- runServerPart Controllers.Course.index
         let actual = BL.unpack $ rsBody response
         assertEqual ("Unexpected response body for " ++ label) expected actual