+++ app/Tests/Main.hs
@@ -5,28 +5,23 @@ Module that acts as interface for testing multiple test suites using cabal.
 
 -}
 
-module Main
-(  main  ) where
+module Main where
 
 import Control.Monad
 import Config (databasePath)
 import Data.Text (unpack)
 import Database.Database(setupDatabase)
-import RequirementTests.ModifierTests (modifierTestSuite)
-import RequirementTests.PostParserTests (postTestSuite)
-import RequirementTests.PreProcessingTests (preProcTestSuite)
-import RequirementTests.ReqParserTests (reqTestSuite)
-import RequirementTests.CourseControllerTests (courseContTestSuite)
+import RequirementTests.RequirementTests (requirementTests)
 import System.Directory (removeFile)
 import System.Environment (setEnv, unsetEnv)
 import qualified System.Exit as Exit
 import Test.HUnit (Test (..), failures, runTestTT)
+import Tests.CourseControllerTests (courseContTestSuite)
 
--- Single test encompassing all test suites
 tests :: IO Test
 tests = do
     courseTest <- courseContTestSuite
-    return $ TestList [reqTestSuite, postTestSuite, preProcTestSuite, modifierTestSuite, courseTest]
+    return $ TestList [requirementTests, courseTest]
 
 main :: IO ()
 main = do